Extract `FakeComposer` class for re-use.
